+++ src/richie/apps/courses/models/program.py
@@ -2,6 +2,7 @@
 Declare and configure the models for the program part
 """
 
+from django.core.validators import MinValueValidator
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
@@ -9,8 +10,12 @@ from cms.api import Page
 from cms.extensions.extension_pool import extension_pool
 from cms.models.pluginmodel import CMSPlugin
 
+from ...core.fields.duration import CompositeDurationField
 from ...core.models import BasePageExtension
-from ..defaults import PROGRAMS_PAGE
+from .. import defaults
+from .category import Category, CategoryPluginModel
+from .organization import Organization, OrganizationPluginModel
+from .person import Person, PersonPluginModel
 
 
 class Program(BasePageExtension):
@@ -18,7 +23,36 @@ class Program(BasePageExtension):
     The program extension represents and records a program.
     """
 
-    PAGE = PROGRAMS_PAGE
+    PAGE = defaults.PROGRAMS_PAGE
+
+    duration = CompositeDurationField(
+        time_units=defaults.TIME_UNITS,
+        default_unit=defaults.DEFAULT_TIME_UNIT,
+        max_length=80,
+        blank=True,
+        null=True,
+        help_text=_("The program time range."),
+    )
+
+    effort = CompositeDurationField(
+        time_units=defaults.EFFORT_UNITS,
+        default_unit=defaults.DEFAULT_EFFORT_UNIT,
+        max_length=80,
+        blank=True,
+        null=True,
+        help_text=_("Total amount of time to complete this program."),
+    )
+
+    price = models.DecimalField(
+        _("price"),
+        max_digits=10,
+        decimal_places=2,
+        default=0,
+        blank=True,
+        null=True,
+        help_text=_("The price of the program."),
+        validators=[MinValueValidator(0)],
+    )
 
     class Meta:
         db_table = "richie_program"
@@ -31,6 +65,61 @@ class Program(BasePageExtension):
         model = self._meta.verbose_name.title()
         name = self.extended_object.get_title()
         return f"{model:s}: {name:s}"
+
+    @property
+    def pt_effort(self):
+        """Return effort as a PT string for schema.org metadata."""
+        if not self.effort:
+            return ""
+
+        (effort, effort_unit) = self.effort
+        unit_letter = effort_unit[0].upper()
+        return f"PT{effort:d}{unit_letter:s}"
+
+    @property
+    def price_with_currency(self):
+        """Return price with currency for schema.org metadata."""
+        if not self.price:
+            return ""
+
+        return f"€{self.price}"
+
+    def get_categories(self, language=None):
+        """
+        Return the categories linked to the program via a category plugin in any of the
+        placeholders on the program detail page, ranked by their `path` to respect the
+        order in the categories tree.
+        """
+        return self.get_direct_related_page_extensions(
+            Category, CategoryPluginModel, language=language
+        )
+
+    def get_organizations(self, language=None):
+        """
+        Return the organizations linked to the course via an organization plugin in any
+        of the placeholders on the course detail page, ranked by their `path` to respect
+        the order in the organizations tree.
+        """
+        return self.get_direct_related_page_extensions(
+            Organization, OrganizationPluginModel, language=language
+        )
+
+    def get_persons(self, language=None):
+        """
+        Return the persons linked to the course via a person plugin in any of the
+        placeholders on the course detail page, ranked by their `path` to respect
+        the order in the persons tree.
+        """
+        return self.get_direct_related_page_extensions(
+            Person, PersonPluginModel, language=language
+        )
+
+    def save(self, *args, **kwargs):
+        """
+        Enforce validation each time an instance is saved
+        """
+        self.full_clean()
+        super().save(*args, **kwargs)
 
 
 class ProgramPluginModel(CMSPlugin):